+++ effectful/CHANGELOG.md
@@ -15,6 +15,8 @@
   ([#237](https://github.com/haskell-effectful/effectful/issues/237)).
 * Add `HasCallStack` constraints where appropriate for better debugging
   experience.
+* Properly roll back changes made to the environment when `OnEmptyKeep` policy
+  for the `NonDet` effect is selected.
 * **Breaking changes**:
   - `localSeqLend`, `localLend`, `localSeqBorrow` and `localBorrow` now take a
     list of effects instead of a single one.